feat: 유틸함수 추가
